Cut the fat, only reference version of Urquhart
