+++ example/expected_output.txt
@@ -1,0 +1,141 @@
+Obs 1 Polygons: 
+0: [66,75] [75,74] [74,66] 
+1: [71,67] [67,65] [65,71] 
+2: [69,68] [68,49] [49,43] [43,59] [59,69] 
+3: [50,44] [44,49] [49,68] [68,75] [75,61] [61,50] 
+4: [44,24] [24,21] [21,15] [15,44] 
+5: [76,52] [52,60] [60,73] [73,76] 
+6: [64,59] [59,69] [69,71] [71,64] 
+7: [45,34] [34,42] [42,70] [70,62] [62,45] 
+8: [72,63] [63,53] [53,62] [62,72] 
+9: [56,48] [48,63] [63,72] [72,74] [74,66] [66,37] [37,56] 
+10: [29,19] [19,21] [21,24] [24,44] [44,50] [50,61] [61,75] [75,66] [66,37] [37,29] 
+11: [20,17] [17,5] [5,1] [1,11] [11,20] 
+12: [7,2] [2,8] [8,23] [23,38] [38,33] [33,22] [22,7] 
+13: [38,59] [59,64] [64,71] [71,65] [65,33] [33,38] 
+14: [58,54] [54,40] [40,17] [17,20] [20,26] [26,47] [47,46] [46,58] 
+15: [70,42] [42,36] [36,55] [55,54] [54,58] [58,70] 
+16: [11,14] [14,25] [25,31] [31,26] [26,20] [20,11] 
+17: [6,14] [14,11] [11,1] [1,0] [0,2] [2,7] [7,6] 
+18: [16,6] [6,14] [14,25] [25,32] [32,30] [30,16] 
+19: [31,25] [25,32] [32,52] [52,76] [76,58] [58,46] [46,47] [47,26] [26,31] 
+20: [15,10] [10,3] [3,12] [12,15] 
+21: [9,18] [18,27] [27,35] [35,53] [53,62] [62,45] [45,34] [34,42] [42,36] [36,55] [55,54] [54,40] [40,17] [17,5] [5,4] [4,9] 
+22: [0,2] [2,3] [3,10] [10,15] [15,21] [21,19] [19,13] [13,18] [18,9] [9,4] [4,5] [5,1] [1,0] 
+23: [13,19] [19,29] [29,37] [37,56] [56,48] [48,63] [63,53] [53,35] [35,27] [27,18] [18,13] 
+24: [8,23] [23,12] [12,3] [3,2] [2,8] 
+25: [44,49] [49,43] [43,59] [59,38] [38,23] [23,12] [12,15] [15,44] 
+26: [60,73] [73,41] [41,30] [30,32] [32,52] [52,60] 
+27: [57,51] [51,28] [28,22] [22,33] [33,65] [65,67] [67,57] 
+28: [28,51] [51,39] [39,41] [41,30] [30,16] [16,6] [6,7] [7,22] [22,28] 
+29: [51,39] [39,41] [41,73] [73,67] [67,57] [57,51] 
+Obs 2 Polygons: 
+0: [66,74] [74,75] [75,66] 
+1: [65,67] [67,71] [71,65] 
+2: [44,15] [15,21] [21,24] [24,44] 
+3: [15,10] [10,3] [3,12] [12,15] 
+4: [60,52] [52,76] [76,73] [73,60] 
+5: [8,23] [23,12] [12,3] [3,2] [2,8] 
+6: [51,39] [39,41] [41,73] [73,67] [67,57] [57,51] 
+7: [69,59] [59,64] [64,71] [71,69] 
+8: [31,25] [25,32] [32,52] [52,76] [76,58] [58,46] [46,47] [47,26] [26,31] 
+9: [71,64] [64,59] [59,38] [38,33] [33,65] [65,71] 
+10: [44,49] [49,68] [68,75] [75,61] [61,50] [50,44] 
+11: [44,49] [49,43] [43,59] [59,38] [38,23] [23,12] [12,15] [15,44] 
+12: [69,68] [68,49] [49,43] [43,59] [59,69] 
+13: [72,63] [63,53] [53,62] [62,72] 
+14: [6,14] [14,11] [11,1] [1,0] [0,2] [2,7] [7,6] 
+15: [32,25] [25,14] [14,6] [6,16] [16,30] [30,32] 
+16: [1,5] [5,17] [17,20] [20,11] [11,1] 
+17: [20,26] [26,47] [47,46] [46,58] [58,54] [54,40] [40,17] [17,20] 
+18: [25,31] [31,26] [26,20] [20,11] [11,14] [14,25] 
+19: [30,41] [41,73] [73,60] [60,52] [52,32] [32,30] 
+20: [28,22] [22,7] [7,6] [6,16] [16,30] [30,41] [41,39] [39,51] [51,28] 
+21: [22,7] [7,2] [2,8] [8,23] [23,38] [38,33] [33,22] 
+22: [28,22] [22,33] [33,65] [65,67] [67,57] [57,51] [51,28] 
+23: [36,42] [42,34] [34,45] [45,62] [62,53] [53,35] [35,27] [27,18] [18,9] [9,4] [4,5] [5,17] [17,40] [40,54] [54,55] [55,36] 
+24: [55,36] [36,42] [42,70] [70,58] [58,54] [54,55] 
+25: [42,34] [34,45] [45,62] [62,70] [70,42] 
+26: [56,37] [37,66] [66,74] [74,72] [72,63] [63,48] [48,56] 
+27: [13,19] [19,29] [29,37] [37,56] [56,48] [48,63] [63,53] [53,35] [35,27] [27,18] [18,13] 
+28: [13,19] [19,21] [21,15] [15,10] [10,3] [3,2] [2,0] [0,1] [1,5] [5,4] [4,9] [9,18] [18,13] 
+29: [50,44] [44,24] [24,21] [21,19] [19,29] [29,37] [37,66] [66,75] [75,61] [61,50] 
+Matching Time: 
+904.368,-14.3401|790.135,440.198
+912.891,-13.6557|797.171,445.056
+907.477,-22.4587|796.89,434.725
+918.497,-18.4262|804.412,443.731
+917.397,-28.0554|808.279,434.844
+912.957,-4.98778|792.89,452.593
+910.244,-2.5735|789.334,453.326
+906.255,7.28491|780.946,459.864
+905.033,20.3051|773.372,470.525
+924.456,8.00196|796.345,469.594
+926.465,6.6479|798.761,469.427
+929.588,1.3674|804.108,466.418
+992.356,-25.3951|871.843,474.662
+984.947,-20.2398|862.849,475.418
+982.138,-28.717|864.659,466.672
+976.178,-40.8563|865.575,453.18
+918.879,-33.559|812.316,430.821
+922.344,-39.3647|818.222,427.529
+951.533,37.7911|804.878,508.935
+953.725,32.8984|809.224,505.796
+964.653,32.5405|818.864,510.955
+966.571,45.1536|814.212,522.835
+950.324,44.7075|800.37,514.318
+929.762,43.9664|782.939,503.385
+932.535,41.5836|786.532,502.71
+940.993,39.6659|794.814,505.283
+933.288,35.4702|790.244,497.794
+928.286,36.2804|785.508,495.993
+913.124,27.4828|776.784,480.788
+925.898,25.3996|788.886,485.377
+911.342,29.8297|774.067,481.928
+932.072,12.8461|800.513,477.599
+927.247,21.1435|792.183,482.368
+966.324,13.2047|829.988,495.052
+952.412,13.4205|817.836,488.276
+955.144,4.32229|824.754,481.766
+972.677,1.16101|841.516,487.805
+966.146,-5.65516|839.272,478.635
+950.808,-15.6265|830.984,462.326
+946.157,-8.27917|823.28,466.359
+941.347,-13.3524|821.655,459.56
+945.455,-24.1138|830.597,452.299
+932.735,-27.4341|821.247,443.058
+945.455,-24.1138|812.803,452.684
+926.379,-26.107|815.08,441.026
+930.242,-14.8746|812.803,452.684
+928.235,-40.2764|823.778,429.688
+988.524,19.053|846.281,511.226
+981.196,27.5371|835.691,514.903
+976.925,22.8184|834.354,508.68
+986.976,10.9622|848.99,503.446
+977.817,-0.518673|846.807,488.923
+988.64,2.52447|854.654,496.974
+966.354,29.6047|821.807,509.265
+978.471,30.6259|831.785,516.213
+975.439,-10.6709|849.828,478.943
+966.595,-8.03354|840.852,476.8
+980.096,-9.40984|853.229,482.366
+958.098,-11.3683|835.165,469.661
+947.995,2.07458|819.69,476.243
+975.503,-18.963|854.034,471.797
+961.282,-17.6976|841.088,465.775
+966.939,-26.4235|850.353,461.051
+933.544,4.4059|806.012,471.029
+937.417,-1.41284|812.278,467.93
+933.385,-4.52209|810.343,463.22
+938.371,13.7486|805.516,481.533
+947.082,16.5557|811.652,488.323
+943.649,20.5586|806.676,490.07
+949.347,13.386|815.2,486.712
+941.956,32.7392|799.115,499.768
+941.124,28.1274|800.703,495.36
+966.953,-31.0843|852.698,457.023
+948.519,-42.6331|842.518,437.799
+957.2,-39.6852|848.559,444.696
+951.407,-30.5865|838.99,449.674
+937.316,-45.5704|834.29,429.649
+964.011,-32.3557|850.787,454.45
