+++ include/urquhart/observation.hpp
@@ -1,9 +1,7 @@
 #pragma once
 
-#include <distance.hpp>
-#include <polygon.hpp>
 #include <tree.hpp>
-#include <map>
+// #include <map>
 
 #include "libqhullcpp/RboxPoints.h"
 #include "libqhullcpp/QhullError.h"
@@ -20,8 +18,23 @@
 namespace urquhart {
     class Observation {
         public:
-            explicit Observation(PointVector& landmarks);
+            explicit Observation(Points& freshLandmarks);
+            explicit Observation(std::vector<std::vector<double>>& freshLandmarks);
             void view();
+
+            // Storage for the positions of all landmarks in this observation (one per column)
+            // All polygons store references to these values
+            Points landmarks;
+
+            // Storage for each distinct edge in the Delaunay triangulation of this observation
+            // Each column is a pair of column numbers in "landmarks" 
+            EdgeSet triangulationEdges;
+            Eigen::VectorXd triangulationEdgeLengths; // indices should match triangulationEdges
+            
+            // cantor pair of vertex IDs -> EdgeSet column ID (which contains the "landmarks" indices for the two points in this edge)
+            // useful when preventing duplicate operations on 
+            std::unordered_map<size_t, int> edgeRefMap;
+
             // H stores all polygons in a tree structure, where each vertex represents a polygon.
             // The childs of a polygon are the triangles that were merged to compose it.
             // Triangles are the leaves of the tree.
@@ -29,16 +42,15 @@ namespace urquhart {
             // TODO: H should be private and have accessors
             Tree* H;
         private:
-            PointVector landmarks;
             // Computes a Delaunay triangulation using QHull from a set of landmarks.
-            void delaunayTriangulation_(PointVector& points, std::vector<Polygon>& polygons);
-            // Uses the triangles of the delaunay triangulation to compute an urquhart tessellation
+            void delaunayTriangulationFromScratch(std::vector<Polygon>& polygons);
+            
+            // Uses the triangles of the delaunay triangulation to build an "urquhart tessellation"
             void urquhartTesselation_();
             
-            std::vector<EdgeT>::iterator findEdge_(Polygon& x, EdgeT commonEdge);
-            // Merges two polygons into a new one, this is used inside the urquhart tessellation computation
-            // p will be merged with n, which shares the longest edge of p, indexed (relative to p) by commonEdgeIdx
-            Polygon combinePolygonData_(const Polygon& p, const Polygon& n);
-            Polygon mergePolygons_(Polygon& p, Polygon& n, EdgeT commonEdge);
+            // Create a new polygon from existing polyons
+            Polygon mergePolygons_(Polygon& p, Polygon& n, int edgeIndexInP);
+            
+            void printPolygon(const Polygon& p);
     };
 }
