+++ include/urquhart/tree.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <distance.hpp>
 #include <polygon.hpp>
 #include <boost/graph/adjacency_list.hpp>
 #include <boost/graph/breadth_first_search.hpp>
@@ -30,20 +29,26 @@ namespace urquhart {
             explicit Tree(const std::vector<Polygon>& polygons);
             void view_tree();
             void view_h2_polygons();
+
             // Merges two polygons of the tree. If i or j have more than 3 vertices, they are
             // disconnected from the main structure and their children are connected to the new polygon.
             // They are not deleted to keep the incremental indices (this could be improved to reduce memory consumption).
             BoostVertexT merge_op(BoostVertexT i, BoostVertexT j, const Polygon& data);
-            Polygon get_vertex(const BoostVertexT v);
-            // Gets the parent of a polygon
+
+            // Gets the parent of a Polygon in the geometric hierarchy
             BoostVertexT get_ancestor(const BoostVertexT v);
+            Polygon get_vertex(const BoostVertexT v);
+            
             // Gets the children of a vertex. This function reverses the graph and does a search,
             // TODO: this could be simplified by getting the in_edges of v which should be always 1
+            // ^^ Unclear whether this was resolved 
             std::vector<BoostVertexT> get_children(const BoostVertexT v);
             // Does a BFSearch from the root
             std::vector<BoostVertexT> traverse();
             // Does a BFSearch from v
             std::vector<BoostVertexT> traverse(const BoostVertexT v);
+
+
             BoostGraph graph;
             BoostVertexT root;
     };
