+++ src/my_old/poisson_disk_sampling.cpp
